Fix crash involving /reload creaturescripts. (#2635)
Fix crash when we remove a creaturescript (xml or revscriptsys), reloads and execute.

To reproduce this crash, simply create an onThink script for example and register it on a creature, then remove from creaturescripts.xml or delete the file in data/scripts/creaturescripts, then use /reload creaturescripts or /reload scripts. When the creature executes the removed script, the server crashes.