Fix Lua item:decay(decayId) (#2667)
Changed ItemType attribute `decayTo` to ItemAttribute. This attribute is savable to database.
Now it can be customized for every Item.

Previous version of Lua function `item:decay(decayId)` did modify `decayTo` of ItemType. Which did resulted in modifying all items on the server, not just the item on which it was called.

Fixes #2665