require non-zero interval in revscriptsys globalevent think (#3674)
When we create a `globalevent` of type `think` and we forget `globalevent:interval(milliseconds)` before `globalevent:register()`, the script is registered with the interval 0. As this is not checked before, no warning is given and the server runs the script non-stop causing 100% cpu. With this check, we guarantee that the script will not be registered if it does not have a set interval.

The bug can be tested with this simple code:
```
local globalEvent = GlobalEvent('bugThink')
function globalEvent.onTime()
	print('CPU says: "stop, it hurts!"')
	return true
end
-- globalEvent:interval(1000)
globalEvent:register()
```