Encode Latin-1 to UTF-8 and vice versa (#4628)
The issue at hand is that since #2403 (which addressed issue #2175),
all text in the database has been stored using UTF-8 encoding. However,
the data that comes from the client is Latin-1 encoded. This isn't
problematic as long as the characters from the client stay within the
ASCII range, as both Latin-1 and UTF-8 encode these characters the same
way. However, if we receive from the client a string of text containing
characters beyond the ASCII range (e.g., "ñññççç"), this is where data
loss occurs, resulting in the situation described by the original poster
in #4560. We end up storing in the database a set of characters that are
not UTF-8 encoded.

This commit addresses this issue by implementing explicit conversion
mechanisms to handle text encoding conversions seamlessly during data
transmission between the client and server.

The solution at hand may incur too much overhead and should be
evaluated.

* Install Boost.Locale in vcpkg and Docker image

* Remove repeated boost-locale dependency

* Add libiconv as a vcpkg dependency

This is required in MacOS to link against Boost::locale... MacOS, am I
right? :-)

---------

Co-authored-by: Ranieri Althoff <ranisalt@gmail.com>